[PyTorch] Put functor at the end of wrap_kernel_functor_unboxed params
In the (hopefully) common case where the unboxed functor is
really a `c10::CompileTimeFunctionPointer`, this enables a jump
straight to said functor.

Note that I had no end of trouble getting this to have decent
performance without the previous diff, and even with it, all the
manual C10_ALWAYS_INLINE I added were necessary.

Differential Revision: [D26069629](https://our.internmc.facebook.com/intern/diff/D26069629/)

ghstack-source-id: 120446691
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51164