[PyTorch] Save a single add instruction in the dispatcher
This saves one (1) add instruction. New code comments should
explain exactly why. In short, we store a direct pointer in
`OperatorHandle` in addition to the `std::list<OperatorDef>::iterator`
because converting the latter to the former requires an add instruction.

It is not clear to me whether this is a particularly great tradeoff,
but I spent (more) time on it (than I expected), so here it is for
review.

Differential Revision: [D26564192](https://our.internmc.facebook.com/intern/diff/D26564192/)

ghstack-source-id: 122147199
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52543