ns for fx: remove subgraphs from user facing API
Summary:

Changes the user facing API of `prepare_single_model_output` to
require a list of nodes instead of a list of subgraphs. This ensures
that how we define a subgraph is an implementation detail and is
not exposed to the user, keeping the eng cost of updating this
implementation later low.

Test Plan:

```
python test/test_quantization.py TestFXNumericSuiteCoreAPIs
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 0f7ab6dd744b59222e714f16248456d63e281591
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52928